Refactor - Introduced Layout component to streamline structure and removed sidebar and theme management from individual pages
