+++ src/pages/Themes.jsx
@@ -1,12 +1,9 @@
-import Header from "../components/Header";
-import Sidebar from "../components/Sidebar";
+import Layout from "../components/Layout";
 import { useState, useEffect } from "react";
-import { Sun } from "lucide-react";
 
 function Themes() {
-  const [sidebarVisible, setSidebarVisible] = useState(false);
   const themes = [
-    { value: "light", label: "Light", icon: "<Sun />" },
+    { value: "light", label: "Light" },
     { value: "dark", label: "Dark" },
     { value: "navy", label: "Navy" },
     { value: "beartheme", label: "Bear" },
@@ -36,16 +33,10 @@ function Themes() {
     localStorage.setItem("theme", theme);
   }, [theme]);
 
-  function toggleSidebar() {
-    setSidebarVisible(!sidebarVisible);
-  }
-
   return (
-    <div className="w-screen h-screen bg-bg">
-      <Header toggleSidebar={toggleSidebar} sidebarVisible={sidebarVisible} />
-      <Sidebar active={sidebarVisible} />
-      <main className="flex justify-center p-6">
-        <section className="w-[500px] space-y-4">
+    <>
+      <Layout>
+        <section className="w-[500px] space-y-4 ">
           <div className="space-y-2 p-6 bg-bg-card rounded-md shadow text-xl">
             <h3 className="text-2xl text-center font-bold text-tx-title mb-5">
               Select Theme
@@ -53,11 +44,11 @@ function Themes() {
             {themes.map((item) => (
               <div
                 key={item.value}
-                className="flex items-center bg-bg-comp tx-tx-card p-2 rounded-md hover:bg-bg-comp-hover hover:cursor-pointer transition-all duration-200"
+                className="flex items-center p-2 rounded-md transition-all duration-200 bg-bg-comp tx-tx-card hover:bg-bg-comp-hover hover:cursor-pointer"
                 onClick={() => setTheme(item.value)}
               >
                 <input
-                  className="h-4 w-4 cursor-pointer ml-2 accent-bg-comp"
+                  className="h-4 w-4 cursor-pointer ml-2 accent-bg-comp-hover"
                   type="radio"
                   name="theme"
                   value={item.value}
@@ -75,11 +66,8 @@ function Themes() {
             ))}
           </div>
         </section>
-      </main>
-      <footer className="fixed bottom-0 w-full text-tx-header bg-bg-header p-2 font-bold">
-        &copy; 2025 The Checklist Pal
-      </footer>
-    </div>
+      </Layout>
+    </>
   );
 }
